+++ changelog.md
@@ -1,3 +1,8 @@
+## v5.2.0
+
++ Conditionally publish and load database migration files to prevent conflicts in table names
++ Removed class names from database migrations to prevent conflicts
+
 ## v5.1.0
 
 + Update dependencies to support Laravel 10